requirements.txt
pydantic>=2.0
openai>=1.0
python-dotenv>=1.0
tqdm>=4.0
textual>=0.40.0
